backend/requirements.txt
transformers
accelerate
torch
torchvision
pdf2image
pillow
requests
